refactor#61: callback 분리
